Scan fixes batch 3: status/diagnostics propagation
- SVD.singularValues returns the inner bidiagonal QR SVDInfo (was a redundant
  int k); Analysis.cond and Blas.matrixL2 return NaN on non-convergence,
  Analysis.rank throws InvalidOperationException — none of them read the
  unwritten S buffer anymore.
- BlockJacobi/ILU0/IC0 gain non-throwing out-DirectSolveInfo constructor and
  arena-wrapper overloads; BlockJacobi propagates the failing block LU status
  it previously discarded; failed arena builds free their record slot. The
  info-less signatures remain as throwing conveniences delegating to the
  status path.
- LOBPCGInfo doc example matches actual ToFixedString output; SVDInfo.converged
  contract covers the truncated/randomized reduced-panel paths;
  fProxyFullStats.count typed int (float variant lost exact counts past 2^24).
- Tests: singular-diagonal-block status path + throwing convenience, and a
  success-path test driving all three status overloads.

Suite: 6046/6046 green.

+++ Assets/LinearAlgebra/Source/Statistics/Structs.double.cs
@@ -25,7 +25,7 @@ namespace LinearAlgebra
     [BurstCompile]
     public struct doubleFullStats
     {
-        public double count;
+        public int count;
         public double mean;
         public double min;
         public double max;
@@ -37,7 +37,7 @@ namespace LinearAlgebra
         public double q1;
         public double q3;
 
-        public doubleFullStats(double count, double mean, double min, double max, double range, double median, double stdDev, double variance, double iqr, double q1, double q3)
+        public doubleFullStats(int count, double mean, double min, double max, double range, double median, double stdDev, double variance, double iqr, double q1, double q3)
         {
             this.count = count;
             this.mean = mean;
